채팅 메시지 리팩토링, 결제 시스템 구현

+++ src/main/java/kwh/cofshop/coupon/domain/Coupon.java
@@ -105,7 +105,7 @@ public class Coupon {
         };
     }
 
-    public int calculateDiscount(int targetPrice) {
+    public Long calculateDiscount(Long targetPrice) {
         return getPolicy().calculateDiscount(targetPrice);
     }
