+++ src/main/java/kwh/cofshop/coupon/policy/CouponDiscountPolicy.java
@@ -1,5 +1,5 @@
 package kwh.cofshop.coupon.policy;
 
 public interface CouponDiscountPolicy {
-    int calculateDiscount(int value);
+    Long calculateDiscount(Long value);
 }
